large test helper refactor (#4749)
drive collection tests needed a facelift for readability and ease of
production/consumption. This change is primarily housecleaning with
those priorities in mind. It also contains the support for multi-delta
enumerator mocks so that the next PR can focus entirely on testing the
multi-delta enum changes.
