+++ src/internal/m365/collection/drive/collections_tree_test.go
@@ -9,261 +9,19 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/alcionai/corso/src/internal/common/idname"
 	"github.com/alcionai/corso/src/internal/common/prefixmatcher"
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/data"
-	dataMock "github.com/alcionai/corso/src/internal/data/mock"
 	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
 	"github.com/alcionai/corso/src/internal/m365/service/onedrive/mock"
-	"github.com/alcionai/corso/src/internal/m365/support"
 	"github.com/alcionai/corso/src/internal/tester"
-	bupMD "github.com/alcionai/corso/src/pkg/backup/metadata"
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/count"
 	countTD "github.com/alcionai/corso/src/pkg/count/testdata"
 	"github.com/alcionai/corso/src/pkg/fault"
-	"github.com/alcionai/corso/src/pkg/services/m365/api/graph"
 	apiMock "github.com/alcionai/corso/src/pkg/services/m365/api/mock"
 	"github.com/alcionai/corso/src/pkg/services/m365/api/pagers"
 )
-
-// ---------------------------------------------------------------------------
-// helpers
-// ---------------------------------------------------------------------------
-
-func collWithMBH(mbh BackupHandler) *Collections {
-	return NewCollections(
-		mbh,
-		tenant,
-		idname.NewProvider(user, user),
-		func(*support.ControllerOperationStatus) {},
-		control.Options{ToggleFeatures: control.Toggles{
-			UseDeltaTree: true,
-		}},
-		count.New())
-}
-
-func collWithMBHAndOpts(
-	mbh BackupHandler,
-	opts control.Options,
-) *Collections {
-	return NewCollections(
-		mbh,
-		tenant,
-		idname.NewProvider(user, user),
-		func(*support.ControllerOperationStatus) {},
-		opts,
-		count.New())
-}
-
-// func fullOrPrevPath(
-// 	t *testing.T,
-// 	coll data.BackupCollection,
-// ) path.Path {
-// 	var collPath path.Path
-
-// 	if coll.State() != data.DeletedState {
-// 		collPath = coll.FullPath()
-// 	} else {
-// 		collPath = coll.PreviousPath()
-// 	}
-
-// 	require.False(
-// 		t,
-// 		len(collPath.Elements()) < 4,
-// 		"malformed or missing collection path")
-
-// 	return collPath
-// }
-
-func pagerForDrives(drives ...models.Driveable) *apiMock.Pager[models.Driveable] {
-	return &apiMock.Pager[models.Driveable]{
-		ToReturn: []apiMock.PagerResult[models.Driveable]{
-			{Values: drives},
-		},
-	}
-}
-
-func makePrevMetadataColls(
-	t *testing.T,
-	mbh BackupHandler,
-	previousPaths map[string]map[string]string,
-) []data.RestoreCollection {
-	pathPrefix, err := mbh.MetadataPathPrefix(tenant)
-	require.NoError(t, err, clues.ToCore(err))
-
-	prevDeltas := map[string]string{}
-
-	for driveID := range previousPaths {
-		prevDeltas[driveID] = idx(delta, "prev")
-	}
-
-	mdColl, err := graph.MakeMetadataCollection(
-		pathPrefix,
-		[]graph.MetadataCollectionEntry{
-			graph.NewMetadataEntry(bupMD.DeltaURLsFileName, prevDeltas),
-			graph.NewMetadataEntry(bupMD.PreviousPathFileName, previousPaths),
-		},
-		func(*support.ControllerOperationStatus) {},
-		count.New())
-	require.NoError(t, err, "creating metadata collection", clues.ToCore(err))
-
-	return []data.RestoreCollection{
-		dataMock.NewUnversionedRestoreCollection(t, data.NoFetchRestoreCollection{Collection: mdColl}),
-	}
-}
-
-// func compareMetadata(
-// 	t *testing.T,
-// 	mdColl data.Collection,
-// 	expectDeltas map[string]string,
-// 	expectPrevPaths map[string]map[string]string,
-// ) {
-// 	ctx, flush := tester.NewContext(t)
-// 	defer flush()
-
-// 	colls := []data.RestoreCollection{
-// 		dataMock.NewUnversionedRestoreCollection(t, data.NoFetchRestoreCollection{Collection: mdColl}),
-// 	}
-
-// 	deltas, prevs, _, err := deserializeAndValidateMetadata(
-// 		ctx,
-// 		colls,
-// 		count.New(),
-// 		fault.New(true))
-// 	require.NoError(t, err, "deserializing metadata", clues.ToCore(err))
-// 	assert.Equal(t, expectDeltas, deltas, "delta urls")
-// 	assert.Equal(t, expectPrevPaths, prevs, "previous paths")
-// }
-
-// for comparisons done by collection state
-type stateAssertion struct {
-	itemIDs []string
-	// should never get set by the user.
-	// this flag gets flipped when calling assertions.compare.
-	// any unseen collection will error on requireNoUnseenCollections
-	// sawCollection bool
-}
-
-// for comparisons done by a given collection path
-type collectionAssertion struct {
-	doNotMerge    assert.BoolAssertionFunc
-	states        map[data.CollectionState]*stateAssertion
-	excludedItems map[string]struct{}
-}
-
-type statesToItemIDs map[data.CollectionState][]string
-
-// TODO(keepers): move excludeItems to a more global position.
-func newCollAssertion(
-	doNotMerge bool,
-	itemsByState statesToItemIDs,
-	excludeItems ...string,
-) collectionAssertion {
-	states := map[data.CollectionState]*stateAssertion{}
-
-	for state, itemIDs := range itemsByState {
-		states[state] = &stateAssertion{
-			itemIDs: itemIDs,
-		}
-	}
-
-	dnm := assert.False
-	if doNotMerge {
-		dnm = assert.True
-	}
-
-	return collectionAssertion{
-		doNotMerge:    dnm,
-		states:        states,
-		excludedItems: makeExcludeMap(excludeItems...),
-	}
-}
-
-// to aggregate all collection-related expectations in the backup
-// map collection path -> collection state -> assertion
-type collectionAssertions map[string]collectionAssertion
-
-// ensure the provided collection matches expectations as set by the test.
-// func (cas collectionAssertions) compare(
-// 	t *testing.T,
-// 	coll data.BackupCollection,
-// 	excludes *prefixmatcher.StringSetMatchBuilder,
-// ) {
-// 	ctx, flush := tester.NewContext(t)
-// 	defer flush()
-
-// 	var (
-// 		itemCh  = coll.Items(ctx, fault.New(true))
-// 		itemIDs = []string{}
-// 	)
-
-// 	p := fullOrPrevPath(t, coll)
-
-// 	for itm := range itemCh {
-// 		itemIDs = append(itemIDs, itm.ID())
-// 	}
-
-// 	expect := cas[p.String()]
-// 	expectState := expect.states[coll.State()]
-// 	expectState.sawCollection = true
-
-// 	assert.ElementsMatchf(
-// 		t,
-// 		expectState.itemIDs,
-// 		itemIDs,
-// 		"expected all items to match in collection with:\nstate %q\npath %q",
-// 		coll.State(),
-// 		p)
-
-// 	expect.doNotMerge(
-// 		t,
-// 		coll.DoNotMergeItems(),
-// 		"expected collection to have the appropariate doNotMerge flag")
-
-// 	if result, ok := excludes.Get(p.String()); ok {
-// 		assert.Equal(
-// 			t,
-// 			expect.excludedItems,
-// 			result,
-// 			"excluded items")
-// 	}
-// }
-
-// ensure that no collections in the expected set are still flagged
-// as sawCollection == false.
-// func (cas collectionAssertions) requireNoUnseenCollections(
-// 	t *testing.T,
-// ) {
-// 	for p, withPath := range cas {
-// 		for _, state := range withPath.states {
-// 			require.True(
-// 				t,
-// 				state.sawCollection,
-// 				"results should have contained collection:\n\t%q\t\n%q",
-// 				state, p)
-// 		}
-// 	}
-// }
-
-func pageItems(items ...models.DriveItemable) []models.DriveItemable {
-	return append([]models.DriveItemable{driveRootItem()}, items...)
-}
-
-func pagesOf(pages ...[]models.DriveItemable) []mock.NextPage {
-	mnp := []mock.NextPage{}
-
-	for _, page := range pages {
-		mnp = append(mnp, mock.NextPage{Items: page})
-	}
-
-	return mnp
-}
-
-// ---------------------------------------------------------------------------
-// tests
-// ---------------------------------------------------------------------------
 
 type CollectionsTreeUnitSuite struct {
 	tester.Suite
@@ -404,7 +162,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_GetTree() {
 	table := []struct {
 		name          string
 		drivePager    *apiMock.Pager[models.Driveable]
-		enumerator    mock.EnumerateItemsDeltaByDrive
+		enumerator    mock.EnumerateDriveItemsDelta
 		previousPaths map[string]map[string]string
 
 		metadata []data.RestoreCollection
@@ -413,14 +171,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_GetTree() {
 		{
 			name:       "not yet implemented",
 			drivePager: pagerForDrives(drv),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages:       pagesOf(pageItems()),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage()))),
 			expect: expected{
 				canUsePrevBackup: assert.False,
 				collAssertions: collectionAssertions{
@@ -433,7 +187,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_GetTree() {
 					count.PrevPaths: 0,
 				},
 				deltas:    map[string]string{},
-				err:       require.Error,
+				err:       require.NoError,
 				prevPaths: map[string]map[string]string{},
 				skips:     0,
 			},
@@ -504,24 +258,14 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_MakeDriveCollections() {
 		name         string
 		drive        models.Driveable
 		drivePager   *apiMock.Pager[models.Driveable]
-		enumerator   mock.EnumerateItemsDeltaByDrive
 		prevPaths    map[string]string
 		expectErr    require.ErrorAssertionFunc
 		expectCounts countTD.Expected
 	}{
 		{
-			name:       "not yet implemented",
-			drive:      drv,
-			drivePager: pagerForDrives(drv),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages:       pagesOf(pageItems()),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
-			expectErr: require.Error,
+			name:      "not yet implemented",
+			drive:     drv,
+			expectErr: require.NoError,
 			expectCounts: countTD.Expected{
 				count.PrevPaths: 0,
 			},
@@ -534,10 +278,16 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_MakeDriveCollections() {
 			ctx, flush := tester.NewContext(t)
 			defer flush()
 
-			mbh := mock.DefaultDriveBHWith(user, test.drivePager, test.enumerator)
+			mbh := mock.DefaultOneDriveBH(user)
+			mbh.DrivePagerV = pagerForDrives(drv)
+			mbh.DriveItemEnumeration = mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage())))
+
 			c := collWithMBH(mbh)
 
-			colls, paths, du, err := c.makeDriveCollections(
+			_, _, _, err := c.makeDriveCollections(
 				ctx,
 				test.drive,
 				test.prevPaths,
@@ -548,11 +298,11 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_MakeDriveCollections() {
 
 			// TODO(keepers): awaiting implementation
 			test.expectErr(t, err, clues.ToCore(err))
-			assert.Empty(t, colls)
-			assert.Empty(t, paths)
-			assert.Equal(t, id(delta), du.URL)
+			// assert.Empty(t, colls)
+			// assert.Empty(t, paths)
+			// assert.Empty(t, delta.URL)
 
-			test.expectCounts.Compare(t, c.counter)
+			// test.expectCounts.Compare(t, c.counter)
 		})
 	}
 }
@@ -571,7 +321,7 @@ type populateTreeExpected struct {
 
 type populateTreeTest struct {
 	name       string
-	enumerator mock.EnumerateItemsDeltaByDrive
+	enumerator mock.EnumerateDriveItemsDelta
 	tree       *folderyMcFolderFace
 	limiter    *pagerLimiter
 	expect     populateTreeExpected
@@ -588,11 +338,16 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "nil page",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
+			// special case enumerator to generate a null page.
+			// otherwise all enumerators should be DriveEnumerator()s.
+			enumerator: mock.EnumerateDriveItemsDelta{
+				DrivePagers: map[string]*mock.DriveDeltaEnumerator{
 					id(drive): {
-						Pages:       nil,
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
+						DriveID: id(drive),
+						DeltaQueries: []*mock.DeltaQuery{{
+							Pages:       nil,
+							DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
+						}},
 					},
 				},
 			},
@@ -611,14 +366,11 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "root only",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages:       pagesOf(pageItems()),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(),
+					))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -640,14 +392,12 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "root only on two pages",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages:       pagesOf(pageItems(), pageItems()),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(),
+						aPage(),
+					))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -669,19 +419,15 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "many folders in a hierarchy across multiple pages",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(driveItem(id(folder), name(folder), parentDir(), rootID, isFolder)),
-							pageItems(driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder)),
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(), id(folder), isFolder))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(folderAtRoot()),
+						aPage(folderxAtRoot("sib")),
+						aPage(
+							folderAtRoot(),
+							folderxAt("chld", folder),
+						)))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -706,24 +452,20 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "many folders with files",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)),
-							pageItems(
-								driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder),
-								driveItem(idx(file, "sib"), namex(file, "sib"), parentDir(namex(folder, "sib")), idx(folder, "sib"), isFile)),
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(), id(folder), isFolder),
-								driveItem(idx(file, "chld"), namex(file, "chld"), parentDir(namex(folder, "chld")), idx(folder, "chld"), isFile))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							fileAt(folder)),
+						aPage(
+							folderxAtRoot("sib"),
+							filexAt("fsib", "sib")),
+						aPage(
+							folderAtRoot(),
+							folderxAt("chld", folder),
+							filexAt("fchld", "chld")),
+					))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -749,65 +491,60 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 				},
 			},
 		},
-		// TODO: restore after mock.DriveEnumerator support lands.
-		// {
-		// 	name: "many folders with files across multiple deltas",
-		// 	tree: newFolderyMcFolderFace(nil, rootID),
-		// 	enumerator: mock.DriveEnumerator(
-		// 		mock.Drive(id(drive)).With(
-		// 			mock.Delta(id(delta), nil).With(aPage(
-		// 				folderAtRoot(),
-		// 				fileAt(folder))),
-		// 			mock.Delta(id(delta), nil).With(aPage(
-		// 				folderxAtRoot("sib"),
-		// 				filexAt("fsib", "sib"))),
-		// 			mock.Delta(id(delta), nil).With(aPage(
-		// 				folderAtRoot(),
-		// 				folderxAt("chld", folder),
-		// 				filexAt("fchld", "chld"))),
-		// 		)),
-		// 	limiter: newPagerLimiter(control.DefaultOptions()),
-		// 	expect: populateTreeExpected{
-		// 		counts: countTD.Expected{
-		// 			count.TotalFoldersProcessed: 7,
-		// 			count.TotalFilesProcessed:   3,
-		// 			count.TotalPagesEnumerated:  4,
-		// 		},
-		// 		err:            require.NoError,
-		// 		numLiveFiles:   3,
-		// 		numLiveFolders: 4,
-		// 		sizeBytes:      3 * 42,
-		// 		treeContainsFolderIDs: []string{
-		// 			rootID,
-		// 			id(folder),
-		// 			idx(folder, "sib"),
-		// 			idx(folder, "chld"),
-		// 		},
-		// 		treeContainsTombstoneIDs: []string{},
-		// 		treeContainsFileIDsWithParent: map[string]string{
-		// 			id(file):           id(folder),
-		// 			idx(file, "fsib"):  idx(folder, "sib"),
-		// 			idx(file, "fchld"): idx(folder, "chld"),
-		// 		},
-		// 	},
-		// },
+		{
+			name: "many folders with files across multiple deltas",
+			tree: newFolderyMcFolderFace(nil, rootID),
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(aPage(
+						folderAtRoot(),
+						fileAt(folder))),
+					mock.Delta(id(delta), nil).With(aPage(
+						folderxAtRoot("sib"),
+						filexAt("fsib", "sib"))),
+					mock.Delta(id(delta), nil).With(aPage(
+						folderAtRoot(),
+						folderxAt("chld", folder),
+						filexAt("fchld", "chld"))),
+				)),
+			limiter: newPagerLimiter(control.DefaultOptions()),
+			expect: populateTreeExpected{
+				counts: countTD.Expected{
+					count.TotalFoldersProcessed: 7,
+					count.TotalFilesProcessed:   3,
+					count.TotalPagesEnumerated:  4,
+				},
+				err:            require.NoError,
+				numLiveFiles:   3,
+				numLiveFolders: 4,
+				sizeBytes:      3 * 42,
+				treeContainsFolderIDs: []string{
+					rootID,
+					id(folder),
+					idx(folder, "sib"),
+					idx(folder, "chld"),
+				},
+				treeContainsTombstoneIDs: []string{},
+				treeContainsFileIDsWithParent: map[string]string{
+					id(file):           id(folder),
+					idx(file, "fsib"):  idx(folder, "sib"),
+					idx(file, "fchld"): idx(folder, "chld"),
+				},
+			},
+		},
 		{
 			// technically you won't see this behavior from graph deltas, since deletes always
 			// precede creates/updates.  But it's worth checking that we can handle it anyways.
 			name: "create, delete on next page",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)),
-							pageItems(delItem(id(folder), parentDir(), rootID, isFolder))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							fileAt(folder)),
+						aPage(delItem(id(folder), rootID, isFolder)),
+					))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -834,19 +571,15 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 			// precede creates/updates.  But it's worth checking that we can handle it anyways.
 			name: "move->delete folder with populated tree",
 			tree: treeWithFolders(),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(
-								driveItem(idx(folder, "parent"), namex(folder, "parent"), parentDir(), rootID, isFolder),
-								driveItem(id(folder), namex(folder, "moved"), parentDir(), idx(folder, "parent"), isFolder),
-								driveItem(id(file), name(file), parentDir(namex(folder, "parent"), name(folder)), id(folder), isFile)),
-							pageItems(delItem(id(folder), parentDir(), idx(folder, "parent"), isFolder))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderxAtRoot("parent"),
+							driveItem(id(folder), namex(folder, "moved"), parentDir(), idx(folder, "parent"), isFolder),
+							fileAtDeep(parentDir(namex(folder, "parent"), name(folder)), id(folder))),
+						aPage(delItem(id(folder), idx(folder, "parent"), isFolder)),
+					))),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -874,24 +607,20 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "at folder limit before enumeration",
 			tree: treeWithFileAtRoot(),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)),
-							pageItems(
-								driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder),
-								driveItem(idx(file, "sib"), namex(file, "sib"), parentDir(namex(folder, "sib")), idx(folder, "sib"), isFile)),
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(), id(folder), isFolder),
-								driveItem(idx(file, "chld"), namex(file, "chld"), parentDir(namex(folder, "chld")), idx(folder, "chld"), isFile))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							fileAt(folder)),
+						aPage(
+							folderxAtRoot("sib"),
+							filexAt("fsib", "sib")),
+						aPage(
+							folderAtRoot(),
+							folderxAt("chld", folder),
+							filexAt("fchld", "chld")),
+					))),
 			limiter: newPagerLimiter(minimumLimitOpts()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -915,24 +644,20 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_singleDelta(
 		{
 			name: "hit folder limit during enumeration",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			enumerator: mock.EnumerateItemsDeltaByDrive{
-				DrivePagers: map[string]*mock.DriveItemsDeltaPager{
-					id(drive): {
-						Pages: pagesOf(
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)),
-							pageItems(
-								driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder),
-								driveItem(idx(file, "sib"), namex(file, "sib"), parentDir(namex(folder, "sib")), idx(folder, "sib"), isFile)),
-							pageItems(
-								driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-								driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(), id(folder), isFolder),
-								driveItem(idx(file, "chld"), namex(file, "chld"), parentDir(namex(folder, "chld")), idx(folder, "chld"), isFile))),
-						DeltaUpdate: pagers.DeltaUpdate{URL: id(delta)},
-					},
-				},
-			},
+			enumerator: mock.DriveEnumerator(
+				mock.Drive(id(drive)).With(
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							fileAt(folder)),
+						aPage(
+							folderxAtRoot("sib"),
+							filexAt("fsib", "sib")),
+						aPage(
+							folderAtRoot(),
+							folderxAt("chld", folder),
+							filexAt("fchld", "chld")),
+					))),
 			limiter: newPagerLimiter(minimumLimitOpts()),
 			expect: populateTreeExpected{
 				counts: countTD.Expected{
@@ -984,16 +709,19 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_multiDelta()
 			tree: newFolderyMcFolderFace(nil, rootID),
 			enumerator: mock.DriveEnumerator(
 				mock.Drive(id(drive)).With(
-					mock.Delta(id(delta), nil).With(pagesOf(pageItems(
-						driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-						driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)))...),
-					mock.Delta(id(delta), nil).With(pagesOf(pageItems(
-						driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder),
-						driveItem(idx(file, "fsib"), namex(file, "fsib"), parentDir(namex(folder, "sib")), idx(folder, "sib"), isFolder)))...),
-					mock.Delta(id(delta), nil).With(pagesOf(pageItems(
-						driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-						driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(name(folder)), id(folder), isFolder),
-						driveItem(idx(file, "fchld"), namex(file, "fchld"), parentDir(name(folder), namex(folder, "chld")), idx(folder, "chld"), isFolder)))...),
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							fileAt(folder))),
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderxAtRoot("sib"),
+							filexAt("fsib", "sib"))),
+					mock.Delta(id(delta), nil).With(
+						aPage(
+							folderAtRoot(),
+							folderxAt("chld", folder),
+							filexAt("fchld", "chld"))),
 				)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
@@ -1029,18 +757,21 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_multiDelta()
 			enumerator: mock.DriveEnumerator(
 				mock.Drive(id(drive)).With(
 					mock.Delta(id(delta), nil).
-						With(pagesOf(pageItems(
-							driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-							driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)))...),
+						With(
+							aPage(
+								folderAtRoot(),
+								fileAt(folder))),
 					// a (delete,create) pair in the same delta can occur when
 					// a user deletes and restores an item in-between deltas.
 					mock.Delta(id(delta), nil).
-						With(pagesOf(pageItems(
-							delItem(id(folder), parentDir(), rootID, isFolder),
-							delItem(id(file), parentDir(), id(folder), isFile)))...).
-						With(pagesOf(pageItems(
-							driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-							driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)))...),
+						With(
+							aPage(
+								delItem(id(folder), rootID, isFolder),
+								delItem(id(file), id(folder), isFile))).
+						With(
+							aPage(
+								folderAtRoot(),
+								fileAt(folder))),
 				)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
@@ -1070,13 +801,15 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_multiDelta()
 			enumerator: mock.DriveEnumerator(
 				mock.Drive(id(drive)).With(
 					mock.Delta(id(delta), nil).
-						With(pagesOf(pageItems(
-							driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-							driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)))...),
+						With(
+							aPage(
+								folderAtRoot(),
+								fileAt(folder))),
 					mock.Delta(id(delta), nil).
-						With(pagesOf(pageItems(
-							driveItem(id(folder), namex(folder, "rename"), parentDir(), rootID, isFolder),
-							driveItem(id(file), namex(file, "rename"), parentDir(namex(folder, "rename")), id(folder), isFile)))...),
+						With(
+							aPage(
+								driveItem(id(folder), namex(folder, "rename"), parentDir(), rootID, isFolder),
+								driveItem(id(file), namex(file, "rename"), parentDir(namex(folder, "rename")), id(folder), isFile))),
 				)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
@@ -1109,23 +842,23 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_PopulateTree_multiDelta()
 				mock.Drive(id(drive)).With(
 					mock.Delta(id(delta), nil).With(
 						// first page: create /root/folder and /root/folder/file
-						pagesOf(pageItems(
-							driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-							driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)))...).
+						aPage(
+							folderAtRoot(),
+							fileAt(folder)),
 						// assume the user makes changes at this point:
 						// * create a new /root/folder
 						// * move /root/folder/file from old to new folder (same file ID)
 						// * delete /root/folder
 						// in drive deltas, this will show up as another folder creation sharing
 						// the same dirname, but we won't see the delete until...
-						With(pagesOf(pageItems(
+						aPage(
 							driveItem(idx(folder, 2), name(folder), parentDir(), rootID, isFolder),
-							driveItem(id(file), name(file), parentDir(name(folder)), idx(folder, 2), isFile)))...),
+							driveItem(id(file), name(file), parentDir(name(folder)), idx(folder, 2), isFile))),
 					// the next delta, containing the delete marker for the original /root/folder
 					mock.Delta(id(delta), nil).
-						With(pagesOf(pageItems(
-							delItem(id(folder), parentDir(), rootID, isFolder),
-						))...),
+						With(aPage(
+							delItem(id(folder), rootID, isFolder),
+						)),
 				)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: populateTreeExpected{
@@ -1223,7 +956,7 @@ func runPopulateTreeTestMulti(
 	ctx, flush := tester.NewContext(t)
 	defer flush()
 
-	mbh := mock.DefaultDriveBHWithMulti(user, pagerForDrives(drv), test.enumerator)
+	mbh := mock.DefaultDriveBHWith(user, pagerForDrives(drv), test.enumerator)
 	c := collWithMBH(mbh)
 	counter := count.New()
 
@@ -1294,14 +1027,14 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 	table := []struct {
 		name    string
 		tree    *folderyMcFolderFace
-		page    []models.DriveItemable
+		page    mock.NextPage
 		limiter *pagerLimiter
 		expect  expected
 	}{
 		{
 			name:    "nil page",
 			tree:    treeWithRoot(),
-			page:    nil,
+			page:    mock.NextPage{},
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts:   countTD.Expected{},
@@ -1316,7 +1049,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name:    "empty page",
 			tree:    treeWithRoot(),
-			page:    []models.DriveItemable{},
+			page:    mock.NextPage{Items: []models.DriveItemable{}},
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts:   countTD.Expected{},
@@ -1331,7 +1064,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name:    "root only",
 			tree:    treeWithRoot(),
-			page:    pageItems(),
+			page:    aPage(),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1348,10 +1081,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name: "many folders in a hierarchy",
 			tree: treeWithRoot(),
-			page: pageItems(
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-				driveItem(idx(folder, "sib"), namex(folder, "sib"), parentDir(), rootID, isFolder),
-				driveItem(idx(folder, "chld"), namex(folder, "chld"), parentDir(name(folder)), id(folder), isFolder)),
+			page: aPage(
+				folderAtRoot(),
+				folderxAtRoot("sib"),
+				folderxAt("chld", folder)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1371,9 +1104,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name: "create->delete",
 			tree: treeWithRoot(),
-			page: pageItems(
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-				delItem(id(folder), parentDir(), rootID, isFolder)),
+			page: aPage(
+				folderAtRoot(),
+				delItem(id(folder), rootID, isFolder)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1391,10 +1124,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name: "move->delete",
 			tree: treeWithFolders(),
-			page: pageItems(
-				driveItem(idx(folder, "parent"), namex(folder, "parent"), parentDir(), rootID, isFolder),
+			page: aPage(
+				folderxAtRoot("parent"),
 				driveItem(id(folder), namex(folder, "moved"), parentDir(namex(folder, "parent")), idx(folder, "parent"), isFolder),
-				delItem(id(folder), parentDir(), idx(folder, "parent"), isFolder)),
+				delItem(id(folder), idx(folder, "parent"), isFolder)),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1415,9 +1148,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name: "delete->create with previous path",
 			tree: treeWithRoot(),
-			page: pageItems(
-				delItem(id(folder), parentDir(), rootID, isFolder),
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder)),
+			page: aPage(
+				delItem(id(folder), rootID, isFolder),
+				folderAtRoot()),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1436,9 +1169,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 		{
 			name: "delete->create without previous path",
 			tree: treeWithRoot(),
-			page: pageItems(
-				delItem(id(folder), parentDir(), rootID, isFolder),
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder)),
+			page: aPage(
+				delItem(id(folder), rootID, isFolder),
+				folderAtRoot()),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -1469,7 +1202,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_fold
 				ctx,
 				test.tree,
 				drv,
-				test.page,
+				test.page.Items,
 				test.limiter,
 				counter,
 				fault.New(true))
@@ -1502,10 +1235,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFolderToTree() {
 	drv.SetId(ptr.To(id(drive)))
 	drv.SetName(ptr.To(name(drive)))
 
-	fld := driveItem(id(folder), name(folder), parentDir(), rootID, isFolder)
-	subFld := driveItem(id(folder), name(folder), driveParentDir(drv, namex(folder, "parent")), idx(folder, "parent"), isFolder)
+	fld := folderAtRoot()
+	subFld := folderAtDeep(driveParentDir(drv, namex(folder, "parent")), idx(folder, "parent"))
 	pack := driveItem(id(pkg), name(pkg), parentDir(), rootID, isPackage)
-	del := delItem(id(folder), parentDir(), rootID, isFolder)
+	del := delItem(id(folder), rootID, isFolder)
 	mal := malwareItem(idx(folder, "mal"), namex(folder, "mal"), parentDir(), rootID, isFolder)
 
 	type expected struct {
@@ -1776,7 +1509,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_MakeFolderCollectionPath(
 		},
 		{
 			name:      "folder",
-			folder:    driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
+			folder:    folderAtRoot(),
 			expect:    folderPath.String(),
 			expectErr: require.NoError,
 		},
@@ -1828,13 +1561,13 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 	table := []struct {
 		name   string
 		tree   *folderyMcFolderFace
-		page   []models.DriveItemable
+		page   mock.NextPage
 		expect expected
 	}{
 		{
 			name: "one file at root",
 			tree: treeWithRoot(),
-			page: pageItems(driveItem(id(file), name(file), parentDir(name(folder)), rootID, isFile)),
+			page: aPage(fileAtRoot()),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 0,
@@ -1852,9 +1585,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "one file in a folder",
 			tree: newFolderyMcFolderFace(nil, rootID),
-			page: pageItems(
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-				driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile)),
+			page: aPage(
+				folderAtRoot(),
+				fileAt(folder)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 0,
@@ -1872,10 +1605,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "many files in a hierarchy",
 			tree: treeWithRoot(),
-			page: pageItems(
-				driveItem(id(file), name(file), parentDir(), rootID, isFile),
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-				driveItem(idx(file, "chld"), namex(file, "chld"), parentDir(name(folder)), id(folder), isFile)),
+			page: aPage(
+				fileAtRoot(),
+				folderAtRoot(),
+				filexAt("chld", folder)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 0,
@@ -1894,8 +1627,8 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "many updates to the same file",
 			tree: treeWithRoot(),
-			page: pageItems(
-				driveItem(id(file), name(file), parentDir(), rootID, isFile),
+			page: aPage(
+				fileAtRoot(),
 				driveItem(id(file), namex(file, 1), parentDir(), rootID, isFile),
 				driveItem(id(file), namex(file, 2), parentDir(), rootID, isFile)),
 			expect: expected{
@@ -1915,7 +1648,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "delete an existing file",
 			tree: treeWithFileAtRoot(),
-			page: pageItems(delItem(id(file), parentDir(), rootID, isFile)),
+			page: aPage(delItem(id(file), rootID, isFile)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 1,
@@ -1931,9 +1664,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "delete the same file twice",
 			tree: treeWithFileAtRoot(),
-			page: pageItems(
-				delItem(id(file), parentDir(), rootID, isFile),
-				delItem(id(file), parentDir(), rootID, isFile)),
+			page: aPage(
+				delItem(id(file), rootID, isFile),
+				delItem(id(file), rootID, isFile)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 2,
@@ -1949,9 +1682,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "create->delete",
 			tree: treeWithRoot(),
-			page: pageItems(
-				driveItem(id(file), name(file), parentDir(), rootID, isFile),
-				delItem(id(file), parentDir(), rootID, isFile)),
+			page: aPage(
+				fileAtRoot(),
+				delItem(id(file), rootID, isFile)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 1,
@@ -1967,10 +1700,10 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "move->delete",
 			tree: treeWithFileAtRoot(),
-			page: pageItems(
-				driveItem(id(folder), name(folder), parentDir(), rootID, isFolder),
-				driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile),
-				delItem(id(file), parentDir(name(folder)), id(folder), isFile)),
+			page: aPage(
+				folderAtRoot(),
+				fileAt(folder),
+				delItem(id(file), id(folder), isFile)),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 1,
@@ -1986,9 +1719,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "delete->create an existing file",
 			tree: treeWithFileAtRoot(),
-			page: pageItems(
-				delItem(id(file), parentDir(), rootID, isFile),
-				driveItem(id(file), name(file), parentDir(), rootID, isFile)),
+			page: aPage(
+				delItem(id(file), rootID, isFile),
+				fileAtRoot()),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 1,
@@ -2006,9 +1739,9 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 		{
 			name: "delete->create a non-existing file",
 			tree: treeWithRoot(),
-			page: pageItems(
-				delItem(id(file), parentDir(), rootID, isFile),
-				driveItem(id(file), name(file), parentDir(), rootID, isFile)),
+			page: aPage(
+				delItem(id(file), rootID, isFile),
+				fileAtRoot()),
 			expect: expected{
 				counts: countTD.Expected{
 					count.TotalDeleteFilesProcessed: 1,
@@ -2038,7 +1771,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_EnumeratePageOfItems_file
 				ctx,
 				test.tree,
 				drv,
-				test.page,
+				test.page.Items,
 				newPagerLimiter(control.DefaultOptions()),
 				counter,
 				fault.New(true))
@@ -2078,7 +1811,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "add new file",
 			tree:    treeWithRoot(),
-			file:    driveItem(id(file), name(file), parentDir(), rootID, isFile),
+			file:    fileAtRoot(),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2096,7 +1829,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "duplicate file",
 			tree:    treeWithFileAtRoot(),
-			file:    driveItem(id(file), name(file), parentDir(), rootID, isFile),
+			file:    fileAtRoot(),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2114,7 +1847,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "error file seen before parent",
 			tree:    treeWithRoot(),
-			file:    driveItem(id(file), name(file), parentDir(name(folder)), id(folder), isFile),
+			file:    fileAt(folder),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2146,7 +1879,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "delete non-existing file",
 			tree:    treeWithRoot(),
-			file:    delItem(id(file), parentDir(name(folder)), id(folder), isFile),
+			file:    delItem(id(file), id(folder), isFile),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2162,7 +1895,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "delete existing file",
 			tree:    treeWithFileAtRoot(),
-			file:    delItem(id(file), parentDir(), rootID, isFile),
+			file:    delItem(id(file), rootID, isFile),
 			limiter: newPagerLimiter(control.DefaultOptions()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2178,7 +1911,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "already at container file limit",
 			tree:    treeWithFileAtRoot(),
-			file:    driveItem(idx(file, 2), namex(file, 2), parentDir(), rootID, isFile),
+			file:    filexAtRoot(2),
 			limiter: newPagerLimiter(minimumLimitOpts()),
 			expect: expected{
 				counts: countTD.Expected{
@@ -2197,7 +1930,7 @@ func (suite *CollectionsTreeUnitSuite) TestCollections_AddFileToTree() {
 		{
 			name:    "goes over total byte limit",
 			tree:    treeWithRoot(),
-			file:    driveItem(id(file), name(file), parentDir(), rootID, isFile),
+			file:    fileAtRoot(),
 			limiter: newPagerLimiter(minimumLimitOpts()),
 			expect: expected{
 				counts: countTD.Expected{
