+++ src/internal/m365/collection/drive/delta_tree_test.go
@@ -14,75 +14,6 @@ import (
 )
 
 // ---------------------------------------------------------------------------
-// helpers
-// ---------------------------------------------------------------------------
-
-var loc = path.NewElements("root:/foo/bar/baz/qux/fnords/smarf/voi/zumba/bangles/howdyhowdyhowdy")
-
-func treeWithRoot() *folderyMcFolderFace {
-	tree := newFolderyMcFolderFace(nil, rootID)
-	rootey := newNodeyMcNodeFace(nil, rootID, rootName, false)
-	tree.root = rootey
-	tree.folderIDToNode[rootID] = rootey
-
-	return tree
-}
-
-func treeWithTombstone() *folderyMcFolderFace {
-	tree := treeWithRoot()
-	tree.tombstones[id(folder)] = newNodeyMcNodeFace(nil, id(folder), "", false)
-
-	return tree
-}
-
-func treeWithFolders() *folderyMcFolderFace {
-	tree := treeWithRoot()
-
-	parent := newNodeyMcNodeFace(tree.root, idx(folder, "parent"), namex(folder, "parent"), true)
-	tree.folderIDToNode[parent.id] = parent
-	tree.root.children[parent.id] = parent
-
-	f := newNodeyMcNodeFace(parent, id(folder), name(folder), false)
-	tree.folderIDToNode[f.id] = f
-	parent.children[f.id] = f
-
-	return tree
-}
-
-func treeWithFileAtRoot() *folderyMcFolderFace {
-	tree := treeWithRoot()
-	tree.root.files[id(file)] = fileyMcFileFace{
-		lastModified: time.Now(),
-		contentSize:  42,
-	}
-	tree.fileIDToParentID[id(file)] = rootID
-
-	return tree
-}
-
-func treeWithFileInFolder() *folderyMcFolderFace {
-	tree := treeWithFolders()
-	tree.folderIDToNode[id(folder)].files[id(file)] = fileyMcFileFace{
-		lastModified: time.Now(),
-		contentSize:  42,
-	}
-	tree.fileIDToParentID[id(file)] = id(folder)
-
-	return tree
-}
-
-func treeWithFileInTombstone() *folderyMcFolderFace {
-	tree := treeWithTombstone()
-	tree.tombstones[id(folder)].files[id(file)] = fileyMcFileFace{
-		lastModified: time.Now(),
-		contentSize:  42,
-	}
-	tree.fileIDToParentID[id(file)] = id(folder)
-
-	return tree
-}
-
-// ---------------------------------------------------------------------------
 // tests
 // ---------------------------------------------------------------------------
